requirements-dev.txt
-r requirements.txt
factory-boy==2.12.0
flake8==3.8.3
flake8-commas==2.0.0
pytest==6.2.3
pytest-django==4.2.0
twine==2.0.0
